Working version without jest working yet
