+++ package.json
@@ -60,11 +60,9 @@
     "@typescript-eslint/parser": "^6.4.0",
     "eslint": "^8.57.0",
     "eslint-config-love": "^43.1.0",
-    "eslint-plugin-no-for-of-loops": "^1.0.1",
     "eslint-config-prettier": "^9.1.0",
     "eslint-plugin-import": "^2.29.1",
     "eslint-plugin-n": "^16.6.2",
-    "eslint-plugin-no-loops": "^0.3.0",
     "eslint-plugin-node": "^11.1.0",
     "eslint-plugin-prettier": "^5.1.3",
     "eslint-plugin-promise": "^6.1.1",
