+++ src/models/MgmError.ts
@@ -1,8 +1,12 @@
-import { isObject } from '../services/general'
+import { isObject } from '../services/general.js'
 
 /** Custom API error to throw. */
 export default class MgmError<T = unknown> extends Error {
-  constructor(message = 'Error', public statusCode = -1, public response?: T) {
+  constructor(
+    message = 'Error',
+    public statusCode = -1,
+    public response?: T
+  ) {
     super(message)
 
     Object.setPrototypeOf(this, new.target.prototype)
@@ -10,9 +14,7 @@ export default class MgmError<T = unknown> extends Error {
 
   toString() {
     return `${this.message}\nResponse:\n${
-      isObject(this.response)
-        ? JSON.stringify(this.response, null, 2)
-        : this.response ?? ''
+      isObject(this.response) ? JSON.stringify(this.response, null, 2) : this.response ?? ''
     }`
   }
 }
