+++ src/models/id-name.ts
@@ -1,7 +1,7 @@
-import { INameValue } from "./name-value"
-import { IDate, IId, IName, IType } from "./interfaces"
+import { INameValue } from './name-value.js'
+import { IDate, IId, IName, IType } from './interfaces.js'
 
-export interface IIdName<Tid = string, Tname = string> extends IId<Tid>, IName<Tname> { }
+export interface IIdName<Tid = string, Tname = string> extends IId<Tid>, IName<Tname> {}
 
 export class IdName<Tid = string, Tname = string> implements IIdName<Tid, Tname> {
   id: Tid
@@ -13,26 +13,41 @@ export class IdName<Tid = string, Tname = string> implements IIdName<Tid, Tname>
   }
 }
 
-export class IdNameNumber extends IdName<number> { }
+export class IdNameNumber extends IdName<number> {}
 
 export type IdNameType<Tid = string, Tname = string> = {
   id: Tid
   name: Tname
 }
 
-export interface IdNameValue<Tvalue, Tid = string> extends IdName<Tid, string>, INameValue<Tvalue> { }
-export interface IdNameValueType<Tvalue, Type = string, Tid = string> extends IdName<Tid, string>, INameValue<Tvalue>, IType<Type> { }
+export interface IdNameValue<Tvalue, Tid = string>
+  extends IdName<Tid, string>,
+    INameValue<Tvalue> {}
+export interface IdNameValueType<Tvalue, Type = string, Tid = string>
+  extends IdName<Tid, string>,
+    INameValue<Tvalue>,
+    IType<Type> {}
 
 /**
  * Used for sending data to a parent handler in a structured way with contextual id and name.
  */
-export interface IValueChange<Tvalue = string> extends IdNameValueType<Tvalue, string, string>, IDate<number> { }
+export interface IValueChange<Tvalue = string>
+  extends IdNameValueType<Tvalue, string, string>,
+    IDate<number> {}
 /**
  * Used to pass structured data back to a caller. Especially for event handlers.
  */
-export type ValueChangeHandler<Tvalue = string, TReturn = void> = (change: IValueChange<Tvalue>) => TReturn
+export type ValueChangeHandler<Tvalue = string, TReturn = void> = (
+  change: IValueChange<Tvalue>
+) => TReturn
 
-export function createValueChange<T>(id: string, name: string, value: T, type?: string, date?: number) {
+export function createValueChange<T>(
+  id: string,
+  name: string,
+  value: T,
+  type?: string,
+  date?: number
+) {
   const vc: IValueChange<T> = {
     id,
     name,
