+++ src/models/id-val.ts
@@ -1,6 +1,6 @@
-import { IId, IVal } from "./interfaces"
+import { IId, IVal } from './interfaces.js'
 
-export interface IIdVal<Tid = string, Tval = string> extends IId<Tid>, IVal<Tval> { }
+export interface IIdVal<Tid = string, Tval = string> extends IId<Tid>, IVal<Tval> {}
 
 export class IdVal<Tid = string, Tval = string> implements IIdVal<Tid, Tval> {
   id: Tid
