+++ src/models/types.ts
@@ -1,5 +1,5 @@
-import { MgmLoggerLevel } from '../services/MgmLogger'
-import InstrumentationStatistics from './InstrumentationStatistics'
+import { MgmLoggerLevel } from '../services/MgmLogger.js'
+import InstrumentationStatistics from './InstrumentationStatistics.js'
 
 export type MgmApiResponse<T = unknown> = {
   code: number
@@ -25,12 +25,7 @@ export type CreateMutable<Type> = {
   -readonly [Property in keyof Type]: Type[Property]
 }
 
-export type JSONValue =
-  | string
-  | number
-  | boolean
-  | { [x: string]: JSONValue }
-  | JSONValue[]
+export type JSONValue = string | number | boolean | { [x: string]: JSONValue } | JSONValue[]
 
 export type ModifyType<T, R> = Omit<T, keyof R> & R
 export type Opaque<K, T = string> = T & { __TYPE__: K }
