+++ src/services/index.ts
@@ -1,18 +1,16 @@
-export * from './ApiRepository'
-export * from './ApiWrapper'
-export * from './AxiosHelper'
-export { default as AzureHelper } from './AzureHelper'
-export * from './CryptoHelper'
-export * from './DateHelper'
-export * from './general'
-export * from './html-helper'
-export * from './jwt'
-export { default as KeyVault } from './KeyVault'
-export * from './MgmHeaderManager'
-export * from './MgmLogger'
-export * from './number-helper'
-export * from './object-helper'
-export * from './SingleLineFileProcessor'
-export { default as SqlServerHelper } from './SqlServerHelper'
-export { default as TelemetryAppInsights } from './TelemetryAppInsights'
-export * from './WebStorage'
+export * from './ApiRepository.js'
+export * from './ApiWrapper.js'
+export * from './AxiosHelper.js'
+export * from './array-helper.js'
+export * from './CaptureResponse.js'
+export * from './CryptoHelper.js'
+export * from './DateHelper.js'
+export * from './general.js'
+export * from './html-helper.js'
+export * from './jwt.js'
+export * from './MgmLogger.js'
+export * from './number-helper.js'
+export * from './object-helper.js'
+export * from './SingleLineFileProcessor.js'
+export { default as SqlServerHelper } from './SqlServerHelper.js'
+export * from './WebStorage.js'
