+++ src/tplot/DashboardScreenSetting.ts
@@ -3,5 +3,8 @@ import { IDashboardScreenSetting, IGridTileConfig } from './tp-items.js'
 export class DashboardScreenSetting implements IDashboardScreenSetting {
   id = ''
 
-  constructor(public name = '', public tiles: IGridTileConfig[] = []) {}
+  constructor(
+    public name = '',
+    public tiles: IGridTileConfig[] = []
+  ) {}
 }
